refactor: introduce new file structure
