--- components/Auth.astro
@@ -1,15 +1,0 @@
----
-import { type AstroAuthConfig, getSession } from '..'
-
-interface Props {
-  authOpts: AstroAuthConfig
-}
-
-const { authOpts } = Astro.props as Props
-
-let session = await getSession(Astro.request, authOpts)
----
-
-<div>
-  <Fragment set:html={Astro.slots.render('default', [session])} />
-</div>
