+++ src/api/[...auth].ts
@@ -1,3 +1,3 @@
-import { AstroAuth } from "..";
+import { AstroAuth } from "../..";
 
 export const { get, post } = AstroAuth();
