+++ src/components/SignIn.astro
@@ -19,7 +19,7 @@ attrs.class = `sign-in ${attrs.class ?? ''}`
 </button>
 
 <script>
-  import { signIn } from '../client'
+  import { signIn } from '../../client'
   // @ts-ignore
   window.signIn = signIn
 </script>
