+++ src/components/SignOut.astro
@@ -15,7 +15,7 @@ attrs.class = `sign-out ${attrs.class ?? ''}`
 </button>
 
 <script>
-  import { signOut } from '../client'
+  import { signOut } from '../../client'
   // @ts-ignore
   window.signOut = signOut
 </script>
